@@ -1,17 +1,13 @@
 ﻿using NUnit.Framework;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
-namespace EntityFrameworkCore.RawSQLExtensions.Tests.SqlQuery
+namespace EntityFrameworkCore.RawSQLExtensions.Tests.SqlQuery;
+
+[TestFixture]
+public class SqlQueryBaseTests
 {
-    [TestFixture]
-    public class SqlQueryBaseTests
+    [Test]
+    public void Todo()
     {
-        [Test]
-        public void Todo()
-        {
-            // TODO
-        }
+        // TODO
     }
 }
