@@ -1,11 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
+﻿namespace EntityFrameworkCore.RawSQLExtensions.Options;
 
-namespace EntityFrameworkCore.RawSQLExtensions.Options
+public static class RawSQLExtensionsOptions
 {
-    public static class RawSQLExtensionsOptions
-    {
-        public static bool AllowStringGuidConversion { get; set; } = true;
-    }
+    public static bool AllowStringGuidConversion { get; set; } = true;
 }
